Move save bot tx to own struct
