--- bot/main.go
@@ -1,27 +1,0 @@
-package bot
-
-import (
-	"bytes"
-	"encoding/hex"
-	"github.com/jchavannes/btcd/txscript"
-	"github.com/memocash/index/ref/bitcoin/memo"
-)
-
-func grabMessage(outputScripts []string) string {
-	for _, script := range outputScripts {
-		lockScript, err := hex.DecodeString(script)
-		if err != nil {
-			panic(err)
-		}
-		pushData, err := txscript.PushedData(lockScript)
-		if err != nil {
-			panic(err)
-		}
-
-		if len(pushData) > 2 && bytes.Equal(pushData[0], memo.PrefixSendMoney) {
-			message := string(pushData[2])
-			return message
-		}
-	}
-	return ""
-}
